+++ Vulkan/CharacterSceneNode.hpp
@@ -111,73 +111,3 @@ public:
 		}
 	}
 };
-
-//
-//	SceneGraph Create Function
-CharacterSceneNode* SceneGraph::createCharacterSceneNode(const char* FileFBX, btVector3 Position) {
-	Pipeline::Default* Pipe = _Driver->_MaterialCache->GetPipe_Default();
-
-	tinygltf::Model Mdl;
-	_ImportGLTF->loadModel(Mdl, FileFBX);
-	GLTFInfo* Infos = _ImportGLTF->ParseModel(Mdl);
-
-	//	TODO:
-	//	Place this into Import_GLTF
-	std::string DiffuseFile("media/");
-	DiffuseFile += Infos->TexDiffuse;
-	TextureObject* DiffuseTex = Pipe->createTextureImage(DiffuseFile);
-	if (DiffuseTex == nullptr) {
-		return nullptr;
-	}
-	//	END TODO
-
-	TriangleMesh* Mesh = new TriangleMesh(_Driver, Pipe, Infos, DiffuseTex);
-	btCollisionShape* ColShape;
-	if (_CollisionShapes.count(FileFBX) == 0) {
-		DecompResults* Results = Decomp(Infos);
-		ColShape = Results->CompoundShape;
-		_CollisionShapes[FileFBX] = ColShape;
-		for (int i = 0; i < Results->m_convexShapes.size(); i++) {
-			_ConvexShapes.push_back(Results->m_convexShapes[i]);
-		}
-		for (int i = 0; i < Results->m_trimeshes.size(); i++) {
-			_TriangleMeshes.push_back(Results->m_trimeshes[i]);
-		}
-		delete Results;
-	}
-	else {
-		ColShape = _CollisionShapes[FileFBX];
-	}
-
-	CharacterSceneNode* MeshNode = new CharacterSceneNode(Mesh);
-	MeshNode->Name = "Character Scene Node";
-
-	//
-	//	Bullet Physics
-	MeshNode->_CollisionShape = ColShape;
-	btTransform Transform;
-	Transform.setIdentity();
-	Transform.setOrigin(Position);
-	//Transform.setRotation(btQuaternion(btVector3(1, 0, 0), glm::radians(-90.0f)));
-
-	btScalar Mass = 0.5f;
-	bool isDynamic = (Mass != 0.f);
-
-	btVector3 localInertia(0, 0, 0);
-	if (isDynamic) {
-		MeshNode->_CollisionShape->calculateLocalInertia(Mass, localInertia);
-	}
-
-	CharacterSceneNodeMotionState* MotionState = new CharacterSceneNodeMotionState(MeshNode, Transform);
-	btRigidBody::btRigidBodyConstructionInfo rbInfo(Mass, MotionState, MeshNode->_CollisionShape, localInertia);
-	MeshNode->_RigidBody = new btRigidBody(rbInfo);
-	MeshNode->_RigidBody->setUserPointer(MeshNode);
-	MeshNode->_RigidBody->setAngularFactor(btVector3(0.0f, 0.0f, 0.0f));
-	dynamicsWorld->addRigidBody(MeshNode->_RigidBody);
-
-	//
-	//	Push new SceneNode into the SceneGraph
-	SceneNodes.push_back(MeshNode);
-	this->invalidate();
-	return MeshNode;
-}
