+++ Vulkan/SkinnedMeshSceneNode.hpp
@@ -181,70 +181,3 @@ public:
 		}
 	}
 };
-
-//
-//	SceneGraph Create Function
-SkinnedMeshSceneNode* SceneGraph::createSkinnedMeshSceneNode(const char* FileFBX, btScalar Mass, btVector3 Position) {
-	Pipeline::Skinned* Pipe = _Driver->_MaterialCache->GetPipe_Skinned();
-
-
-	tinygltf::Model Mdl;
-	_ImportGLTF->loadModel(Mdl, FileFBX);
-	GLTFInfo* Infos = _ImportGLTF->ParseModel(Mdl);
-
-	//	TODO:
-	//	Place this into Import_GLTF
-	std::string DiffuseFile("media/");
-	DiffuseFile += Infos->TexDiffuse;
-	TextureObject* DiffuseTex = Pipe->createTextureImage(DiffuseFile);
-	if (DiffuseTex == nullptr) {
-		return nullptr;
-	}
-	//	END TODO
-
-	TriangleMesh* Mesh = new TriangleMesh(_Driver, Pipe, Infos, DiffuseTex);
-	btCollisionShape* ColShape;
-	if (_CollisionShapes.count(FileFBX) == 0) {
-		DecompResults* Results = Decomp(Infos);
-		ColShape = Results->CompoundShape;
-		_CollisionShapes[FileFBX] = ColShape;
-		for (int i = 0; i < Results->m_convexShapes.size(); i++) {
-			_ConvexShapes.push_back(Results->m_convexShapes[i]);
-		}
-		for (int i = 0; i < Results->m_trimeshes.size(); i++) {
-			_TriangleMeshes.push_back(Results->m_trimeshes[i]);
-		}
-		delete Results;
-	}
-	else {
-		ColShape = _CollisionShapes[FileFBX];
-	}
-
-	SkinnedMeshSceneNode* MeshNode = new SkinnedMeshSceneNode(Mesh);
-	MeshNode->Name = "SkinnedMeshSceneNode";
-
-	//
-	//	Bullet Physics
-	MeshNode->_CollisionShape = ColShape;
-	btTransform Transform;
-	Transform.setIdentity();
-	Transform.setOrigin(Position);
-	//Transform.setRotation(btQuaternion(btVector3(1, 0, 0), glm::radians(-90.0f)));
-
-	bool isDynamic = (Mass != 0.f);
-
-	btVector3 localInertia(0, 0, 0);
-	if (isDynamic) {
-		MeshNode->_CollisionShape->calculateLocalInertia(Mass, localInertia);
-	}
-
-	SceneNodeMotionState* MotionState = new SceneNodeMotionState(MeshNode, Transform);
-	btRigidBody::btRigidBodyConstructionInfo rbInfo(Mass, MotionState, MeshNode->_CollisionShape, localInertia);
-	MeshNode->_RigidBody = new btRigidBody(rbInfo);
-	MeshNode->_RigidBody->setUserPointer(MeshNode);
-	dynamicsWorld->addRigidBody(MeshNode->_RigidBody);
-
-	SceneNodes.push_back(MeshNode);
-	this->invalidate();
-	return MeshNode;
-}
